feat(runtime): preserve signed64 scalar bindings

+++ src/runtime_cuda.c
@@ -719,9 +719,9 @@ static int poly_cuda_graph_node_update(PolyCudaGraphNode *node, const PolyCudaGr
   for (int i = n_buffer_args; i < n_args; i++) {
     int scalar_idx = i - n_buffer_args;
     if (!args[i]) return -1;
-    /* CUDAGraph uses the same width-aware integer encoding as ordinary
-     * launches. Keep eight bytes alive even for an int32 host binding. */
-    node->scalar_values[scalar_idx] = (int64_t) * (int32_t *)args[i];
+    /* CUDA encode_args copies integer bits into signature-width slots.
+     * Graph replay owns these slots independently of the caller's vals. */
+    node->scalar_values[scalar_idx] = *(int64_t *)args[i];
     node->kernel_params[i] = &node->scalar_values[scalar_idx];
   }
